Use RequireAuth to manage and check auth or redirect, fix load flickering
